Resized Cinder as they were too small, added more brazier functionality (cinematic camera, bird can enter/leave brazier now using a temporary timer setup - this will need reworked to be part of the brazier interface instead of being based on FTimers). Need to get the brazier animated Cinder model, and also create the UI to be used for the brazier storyboards. Added CinematicCamera to the build.cs so we can use it with C++.
First level has all the mechanics currently in our game - hoops, collectibles, and a "working" brazier. Cinder can boost and maintain dive momentum. The moments can now be implemented using the BirdSpline class and Scott's matinee trigger class (may need reworked as matinees are a bit dated). This is the current state of the game's mechanics.