Added a new narrative "book" mechanic as requested by designers. Removed some old classes. Made some quick changes to level two and placed example of narrative spline in level three.
Narrative spline - set points along a spline for a narrative object to follow and rest at until the bird catches up to each resting point. Example is in level three, flaming boat style. Narrator audio cues can be added easily.
BP should be straightforward enough to use - set the points for the book to go to for story points 1, 2 and 3. After player reaching the third point of resting, book will travel to end of spline so maybe have it travel somewhere out of sight/ towards the end of the level.
Be sure to set the spline points that represent the correct places you want the book to rest, in the correct order (1, then 2, then 3). Also remember to set static mesh and particle system.